+++ client_udp_listener.go
@@ -2,7 +2,6 @@ package gortsplib
 
 import (
 	"crypto/rand"
-	"fmt"
 	"math/big"
 	"net"
 	"sync/atomic"
@@ -28,42 +27,6 @@ func randInRange(maxVal int) (int, error) {
 type packetConn interface {
 	net.PacketConn
 	SyscallConn() (syscall.RawConn, error)
-}
-
-func setAndVerifyReadBufferSize(pc packetConn, v int) error {
-	rawConn, err := pc.SyscallConn()
-	if err != nil {
-		panic(err)
-	}
-
-	var err2 error
-
-	err = rawConn.Control(func(fd uintptr) {
-		err2 = syscall.SetsockoptInt(int(fd), syscall.SOL_SOCKET, syscall.SO_RCVBUF, v)
-		if err2 != nil {
-			return
-		}
-
-		var v2 int
-		v2, err2 = syscall.GetsockoptInt(int(fd), syscall.SOL_SOCKET, syscall.SO_RCVBUF)
-		if err2 != nil {
-			return
-		}
-
-		if v2 != (v * 2) {
-			err2 = fmt.Errorf("unable to set read buffer size to %v - check that net.core.rmem_max is greater than %v", v, v)
-			return
-		}
-	})
-	if err != nil {
-		return err
-	}
-
-	if err2 != nil {
-		return err2
-	}
-
-	return nil
 }
 
 type clientUDPListener struct {
